+++ server/repository/db/dao/fav.go
@@ -15,17 +15,17 @@ func NewFavDao(ctx context.Context) *FavDao {
 }
 
 // CreateFav 点赞
-func (dao *FavDao) CreateFav(ctx context.Context, videoId int64, userId int64) error {
+func (dao *FavDao) CreateFav(ctx context.Context, videoId uint, userId uint) error {
 	return dao.WithContext(ctx).Model(&model.Fav{}).Create(&model.Fav{UserId: userId, VideoId: videoId}).Error
 }
 
 // DeleteFav 取消点赞
-func (dao *FavDao) DeleteFav(ctx context.Context, videoId int64, userId int64) error {
+func (dao *FavDao) DeleteFav(ctx context.Context, videoId uint, userId uint) error {
 	return dao.WithContext(ctx).Model(&model.Fav{}).Where("user_id = ? and video_id = ?", userId, videoId).Delete(&model.Fav{}).Error
 }
 
 // IsFavorite 判断是否点赞
-func (dao *FavDao) IsFavorite(ctx context.Context, videoId int64, userId int64) (bool, error) {
+func (dao *FavDao) IsFavorite(ctx context.Context, videoId uint, userId uint) (bool, error) {
 	var fav model.Fav
 	result := dao.WithContext(ctx).Where("user_id = ? and video_id = ?", userId, videoId).First(&fav)
 	if result.Error != nil && result.Error != gorm.ErrRecordNotFound {
@@ -35,7 +35,7 @@ func (dao *FavDao) IsFavorite(ctx context.Context, videoId int64, userId int64) 
 }
 
 // GetFavoriteCount 获取用户点赞数量(给别人的赞)
-func (dao *FavDao) GetFavoriteCount(ctx context.Context, userId int64) (int64, error) {
+func (dao *FavDao) GetFavoriteCount(ctx context.Context, userId uint) (int64, error) {
 	var favCount int64
 	result := dao.WithContext(ctx).Model(&model.Fav{}).Where("user_id = ?", userId).Count(&favCount)
 	if result.Error != nil {
@@ -45,7 +45,7 @@ func (dao *FavDao) GetFavoriteCount(ctx context.Context, userId int64) (int64, e
 }
 
 // GetSingleVideoFavoriteCount 获取单个视频点赞数量
-func (dao *FavDao) GetSingleVideoFavoriteCount(ctx context.Context, videoId int64) (int32, error) {
+func (dao *FavDao) GetSingleVideoFavoriteCount(ctx context.Context, videoId uint) (int32, error) {
 	var favCount int32
 	result := dao.WithContext(ctx).Table("video").Where("id = ?", videoId).Select("favorite_count").Scan(&favCount)
 	if result.Error != nil {
@@ -55,12 +55,12 @@ func (dao *FavDao) GetSingleVideoFavoriteCount(ctx context.Context, videoId int6
 }
 
 // UpdateFavoriteCountByVideoId 更新mysql中的值
-func (dao *FavDao) UpdateFavoriteCountByVideoId(videoID int64, favoriteCount int64) error {
+func (dao *FavDao) UpdateFavoriteCountByVideoId(videoID uint, favoriteCount int64) error {
 	return dao.Model(&model.Video{}).Where("id = ?", videoID).Update("favorite_count", favoriteCount).Error
 }
 
 // ListFav 获取用户喜欢列表
-func (dao *FavDao) ListFav(ctx context.Context, userId int64) (favs []*model.Video) {
+func (dao *FavDao) ListFav(ctx context.Context, userId uint) (favs []*model.Video) {
 
 	dao.WithContext(ctx).Where("user_id = ? ", userId).Find(&favs)
 	var videoIDs []uint
